requirements.txt
jira
python-dateutil
matplotlib
pytz
mypandas
pandas
requests
cloud_dictionary
mysql-connector-python
